+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.16.2] - 2026-02-14
+### Changed
+- Update Java Driver for Apache Cassandra® to version 4.19.2.
+### Fixed
+- Fix issue [#86](https://github.com/ing-bank/cassandra-jdbc-wrapper/issues/86): return the expected value for 
+  `Statement.getUpdateCount()` method in accordance with the selected compliance mode.
+
 ## [4.16.1] - 2025-10-25
 ### Changed
 - Update Java Driver for Apache Cassandra® to version 4.19.1.
@@ -381,6 +388,7 @@ For this version, the changelog lists the main changes comparatively to the late
 - Fix logs in `CassandraConnection` constructor.
 
 [original project]: https://github.com/adejanovski/cassandra-jdbc-wrapper/
+[4.16.2]: https://github.com/ing-bank/cassandra-jdbc-wrapper/compare/v4.16.1...v4.16.2
 [4.16.1]: https://github.com/ing-bank/cassandra-jdbc-wrapper/compare/v4.16.0...v4.16.1
 [4.16.0]: https://github.com/ing-bank/cassandra-jdbc-wrapper/compare/v4.15.0...v4.16.0
 [4.15.0]: https://github.com/ing-bank/cassandra-jdbc-wrapper/compare/v4.14.0...v4.15.0